refactor(connector): split codes and remove unused code
