+++ packages/connector/src/README.md
@@ -1,3 +1,5 @@
 # Connector
 
 A\* path-finder. Based upon https://github.com/bgrins/javascript-astar
+
+Generate map. Based upon https://gist.github.com/jose-mdz/4a8894c152383b9d7a870c24a04447e4
