+++ packages/connector/src/index.ts
@@ -1,3 +1,4 @@
 export * from './a-star.js';
 export * from './create-graph.js';
-export * from './graph.js';
+export * from './rectangle.js';
+export * from './util.js';
